Implement display of cart items in cart screen
Cart items are visible on cart screen if a cart related to the user
is found. Otherwise a different cart screen is shown with options
to return to menu screen or product screen. This commit also changed
the name of the cart_product table to cart_items and now includes a
dml.sql file to help populate tables with dummy data.