Implement creation of carts and cart items
This commit include implementations of creation logic for
carts and cart items in the cart dao and cart service. The
cart item model was also updated to accommodate the information
needed to create a cart_item in the db. Consequently, the
cart item find method was also updated to store the data in the
new cart item model.